Support Haskell extensions properly
Problem: till now, we used a hardcoded set of extensions in the full
interpolator. This means that not every valid code will work in `#{ }`,
and the user has no way to workaround that.

Solution: since `0.8.9`, `haskell-src-meta` has a way to convert
`template-haskell`'s Extension type to `haskell-src-exts`'s `KnownExtension`
type, and this means that know we can pick the extensions available at
the interpolator's call site (via means of `template-haskell`) and pass
them to `parseExp` function.

I did it like: if `haskell-src-meta-0.8.9` or higher is used, then get
the extensions enabled in the module, otherwise use no extensions aside
from the ones provided by the Haskell dialect. If the user needs control
over extensions, we require them to use the recent `haskell-src-meta`
version.
I didn't want to put a `haskell-src-meta >= 0.8.9` boundary, not sure
how soon such a constraint will get justified.