[#21] Asways use LF line endings
Problem:
Currently, when the source file has CRLF line endings, and there is
a newline inside `int` quasiquoter, this quasiquoter creates a line
with `\r`, e.g. `"a\r\nb"`. And if we're writing this line to a file
using `Prelude.writeFile` or `Universum.writeFile`, this string is
automatically converted to `"a\r\r\nb"` when runned  on Windows,
as described in
https://hackage.haskell.org/package/base-4.17.0.0/docs/System-IO.html#g:25

Solution:
As in #21, we should always produce strings with LF line endings.
To do  this, run `T.replace "\r\n" "\n"` on quasiquoter input.