Disable `-Wmissing-kind-signatures`
Problem: with the recent GHC we get a lot of warnings about absent
kind signatures for all our types.

Solution: disable this warning, as the Haskell community seems to agree
that having this warning part of `-Wall` and make it fire on every
simple datatype or typeclass is not good.